adjust storage page to new imports and refactor theme folder location
