+++ src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import { ColorModeScript } from '@chakra-ui/react'
 import { NextPage } from 'next'
 import { Html, Head, Main, NextScript } from 'next/document'
-import { theme } from '../theme/theme'
+import { theme } from '../ui/theme/theme'
 
 const Document: NextPage = () => {
   return (
