+++ src/ui/components/AlertDialogs/index.ts
@@ -1,0 +1,1 @@
+export * from './DeleteItemDialog';
