+++ src/ui/components/Badges/QuantityBadge.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { memo } from 'react'
 import { Badge, BadgeProps } from '@chakra-ui/react'
 
 interface QuantityBadgeProps extends BadgeProps {
@@ -6,20 +6,16 @@ interface QuantityBadgeProps extends BadgeProps {
   storageQuantity: number
 }
 
-const QuantityBadge = ({
-  schemeQuantity,
-  storageQuantity,
-  ...restProps
-}: QuantityBadgeProps) => {
-  let option = { text: 'Dostępne', color: 'green' }
-  if (storageQuantity > 0 && storageQuantity < schemeQuantity)
-    option = { text: 'Częściowo', color: 'yellow' }
-  else if (storageQuantity === 0) option = { text: 'Brak', color: 'red' }
-  return (
-    <Badge variant="solid" colorScheme={option.color} {...restProps}>
-      {option.text}
-    </Badge>
-  )
-}
-
-export default QuantityBadge
+export const QuantityBadge = memo(
+  ({ schemeQuantity, storageQuantity, ...restProps }: QuantityBadgeProps) => {
+    let option = { text: 'Dostępne', color: 'green' }
+    if (storageQuantity > 0 && storageQuantity < schemeQuantity)
+      option = { text: 'Częściowo', color: 'yellow' }
+    else if (storageQuantity === 0) option = { text: 'Brak', color: 'red' }
+    return (
+      <Badge variant="solid" colorScheme={option.color} {...restProps}>
+        {option.text}
+      </Badge>
+    )
+  }
+)
