+++ src/ui/components/Badges/index.ts
@@ -1,0 +1,2 @@
+export * from './RoleBadge';
+export * from './QuantityBadge';
