+++ src/ui/modules/History/DesktopHistory/Filter/Filter.tsx
@@ -1,5 +1,5 @@
 import { Flex, Input, Select, Stack, Text } from '@chakra-ui/react'
-import { useColors } from '../../../../../theme/useColors'
+import { useColors } from '../../../../theme/useColors'
 
 const Filter = () => {
   const colors = useColors()
