+++ src/ui/modules/History/MobileHistory/HistoryListItem/HistoryListItem.tsx
@@ -7,7 +7,7 @@ import {
   Icon,
 } from '@chakra-ui/react'
 import { FaUserCircle } from 'react-icons/fa'
-import { useColors } from '../../../../../theme/useColors'
+import { useColors } from '../../../../theme/useColors'
 import ModalHistory from '../../../../components/Modals/ModalHistory/ModalHistory'
 import NameAndQuantityElement from '../NameAndQuantityElement/NameAndQuantityElement'
