+++ src/ui/modules/List/DesktopList/List/List.tsx
@@ -21,7 +21,7 @@ import { useDispatch } from 'react-redux'
 import { removeExisitngCartList } from '../../../../../store/Slices/storageCartSlice'
 import ListMenu from '../../../../components/Menus/ListMenu'
 import { API_URL } from '../../../../../utils/constants'
-import { useColors } from '../../../../../theme/useColors'
+import { useColors } from '../../../../theme/useColors'
 
 interface Props extends PopulatedCartList {}
