+++ src/ui/modules/Palette/Palette.tsx
@@ -8,7 +8,7 @@ import {
 } from '@chakra-ui/react'
 import DesktopWrapper from '../../components/Wrappers/DesktopWrapper/DesktopWrapper'
 import MobileWrapper from '../../components/Wrappers/MobileWrapper/MobileWrapper'
-import { useColors } from '../../../theme/useColors'
+import { useColors } from '../../theme/useColors'
 import ColorModeSwitch from '../../components/ColorModeSwitch/ColorModeSwitch'
 
 const Rect = (rectProps: BoxProps) => (
@@ -38,7 +38,9 @@ const Palette = () => {
       </HStack>
       <Text as="u">COLORS</Text>
       <Text color={colors.fontPrimary}>FontPrimary</Text>
-      <Text bgColor={colors.fontPrimary} color={colors.fontPrimaryOpposite}>FontPrimaryOpposite</Text>
+      <Text bgColor={colors.fontPrimary} color={colors.fontPrimaryOpposite}>
+        FontPrimaryOpposite
+      </Text>
       <Text color={colors.fontSecondary}>FontSecondary</Text>
       <Text color={colors.fontNeutral}>FontNeutral</Text>
       <Rect bgColor={colors.backgroundPrimary}>BackgroundPrimary</Rect>
