+++ src/ui/modules/Schemes/DesktopSchemes/AddScheme/AddScheme.tsx
@@ -10,14 +10,14 @@ import {
 } from '@chakra-ui/react'
 import { useContext, useState } from 'react'
 import AddItem from './AddItem/AddItem'
-import ProductButton from '../../../../components/Custom Buttons/ProductButton/ProductButton'
+import ProductButton from '../../../../components/CustomButtons/ProductButton/ProductButton'
 import ItemsList from './ItemsList/ItemsList'
 import { useDispatch } from 'react-redux'
 import { fetcher } from '../../../../../utils/requests'
 import { API_URL } from '../../../../../utils/constants'
 import { addSchema } from '../../../../../store/Slices/schemasSlice'
 import { SchemasContext } from '../../../../../pages/schemes'
-import { useColors } from '../../../../../theme/useColors'
+import { useColors } from '../../../../theme/useColors'
 
 const AddScheme = () => {
   const context = useContext(SchemasContext)
