+++ src/ui/modules/Schemes/MobileSchemes/AddScheme/ItemsList/Item/Item.tsx
@@ -13,7 +13,7 @@ import {
 import { AiOutlineClose } from 'react-icons/ai'
 import { TmpSchemaItem } from '../../../../../../../mongo/models/schema'
 import { SchemasContext } from '../../../../../../../pages/schemes'
-import { useColors } from '../../../../../../../theme/useColors'
+import { useColors } from '../../../../../../theme/useColors'
 
 interface ItemProps {
   schemaItem: TmpSchemaItem
