+++ src/ui/modules/Storage/MobileStorage/StorageEdit/CategoryEdit/Category/Category.tsx
@@ -1,7 +1,7 @@
 import { Box, ButtonGroup, Checkbox, Flex, Input } from '@chakra-ui/react'
 import { useState } from 'react'
 import { AiOutlineCheck, AiOutlineClose, AiOutlineEdit } from 'react-icons/ai'
-import { useColors } from '../../../../../../../theme/useColors'
+import { useColors } from 'ui/theme'
 
 interface CategoryProps {
   // FIXME: nie wolno używać key jako propsów
