+++ src/ui/modules/Storage/MobileStorage/StorageEdit/ItemsEdit/ItemsEdit.tsx
@@ -23,8 +23,8 @@ import { useSelector } from 'react-redux'
 import { categoriesInfo } from '../../../../../../store/store'
 import { API_URL } from '../../../../../../utils/constants'
 import { fetcher } from '../../../../../../utils/requests'
-import ProductButton from '../../../../../components/Custom Buttons/ProductButton/ProductButton'
-import { useColors } from '../../../../../../theme/useColors'
+import { ProductButton } from 'ui/components'
+import { useColors } from 'ui/theme'
 
 const ItemsEdit = () => {
   const router = useRouter()
